+++ organization/src/epilot/models/operations/getsettings.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from typing import Any, Optional
-
+from typing import Any, Dict, Optional
 
 
 @dataclasses.dataclass
@@ -12,7 +11,6 @@ class GetSettingsRequest:
     org_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'org_id', 'style': 'simple', 'explode': False }})
     r"""The Id of the organization."""
     
-
 
 
 
@@ -24,7 +22,7 @@ class GetSettingsResponse:
     r"""HTTP response status code for this operation"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
-    settings: Optional[dict[str, Any]] = dataclasses.field(default=None)
+    settings: Optional[Dict[str, Any]] = dataclasses.field(default=None)
     r"""Returns the organization settings"""