+++ organization/src/epilot/models/operations/updateorganization.py
@@ -7,14 +7,12 @@ from ..shared import organization as shared_organization
 from typing import Optional
 
 
-
 @dataclasses.dataclass
 class UpdateOrganizationRequest:
     org_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'org_id', 'style': 'simple', 'explode': False }})
     r"""The Id of the organization."""
     organization: Optional[shared_organization.Organization] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})
     
-