+++ organization/RELEASES.md
@@ -555,3 +555,11 @@ Based on:
 - Speakeasy CLI 1.99.0 (2.154.1) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.47.1] organization
+
+## 2023-10-18 00:27:13
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/organization-v2.yaml
+- Speakeasy CLI 1.101.0 (2.161.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.0] organization